+++ packages/components/src/Chips/InternalChipDismissible/hooks/useInternalChipDismissibleInput.tsx
@@ -6,7 +6,7 @@ import React, {
   useState,
 } from "react";
 import { v1 as uuidV1 } from "uuid";
-import { debounce } from "lodash";
+import debounce from "lodash/debounce";
 import { useLiveAnnounce } from "@jobber/hooks/useLiveAnnounce";
 import {
   ChipDismissibleInputOptionProps,