+++ packages/components/src/List/List.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import classnames from "classnames";
-import { get, groupBy } from "lodash";
+import get from "lodash/get";
+import groupBy from "lodash/groupBy";
 import styles from "./List.css";
 import sectionStyles from "./SectionHeader.css";
 import { ListItem, ListItemProps } from "./ListItem";